55.0.A.2.205: Open Source code

+++ drivers/media/platform/msm/camera/cam_cdm/cam_cdm.h
@@ -195,11 +195,11 @@ struct cam_cdm_hw_intf_cmd_submit_bl {
 	struct cam_cdm_bl_request *data;
 };
 
-/* struct cam_cdm_hw_mem - CDM hw memory struct */
+/* struct cam_cdm_hw_mem - CDM hw memory.struct */
 struct cam_cdm_hw_mem {
 	int32_t handle;
 	uint32_t vaddr;
-	uintptr_t kmdvaddr;
+	uint64_t kmdvaddr;
 	size_t size;
 };
